Fix ILA error in HBM-enabled designs for the 9H7.
When building HBM-enabled designs for the 9H7, an error would occur during opt_design when ILA was disabled.
This was because Vivado automatically adds a dbg_hub component in the HBM IP, which had an unconnected clock port.
This code connects that port to one of the clocks, resolving the error.